@@ -1,5 +1,4 @@
 from collections.abc import Mapping
-from typing import Optional
 
 from pygame.joystick import JoystickType
 
@@ -11,7 +10,7 @@ def get_eventstate() -> bool: ...
 def update() -> None: ...
 def get_count() -> int: ...
 def is_controller(index: int) -> bool: ...
-def name_forindex(index: int) -> Optional[str]: ...
+def name_forindex(index: int) -> str | None: ...
 
 class Controller:
     def __init__(self, index: int) -> None: ...
