@@ -1,5 +1,3 @@
-from typing import Optional
-
 INIT_TIMER: int
 INIT_AUDIO: int
 INIT_VIDEO: int
@@ -11,6 +9,6 @@ INIT_NOPARACHUTE: int
 INIT_EVERYTHING: int
 
 class error(RuntimeError):
-    def __init__(self, message: Optional[str] = None) -> None: ...
+    def __init__(self, message: str | None = None) -> None: ...
 
 def init_subsystem(flags: int) -> None: ...
