@@ -1,4 +1,4 @@
-from typing import Optional, TypedDict
+from typing import TypedDict
 
 class _FingerDict(TypedDict):
     id: int
@@ -9,4 +9,4 @@ class _FingerDict(TypedDict):
 def get_num_devices() -> int: ...
 def get_device(index: int, /) -> int: ...
 def get_num_fingers(device_id: int, /) -> int: ...
-def get_finger(touchid: int, index: int) -> Optional[_FingerDict]: ...
+def get_finger(touchid: int, index: int) -> _FingerDict | None: ...
