@@ -1,6 +1,6 @@
 import sys
 from collections.abc import Collection, Iterator
-from typing import Any, ClassVar, SupportsIndex, Union, overload
+from typing import Any, ClassVar, SupportsIndex, overload
 
 from pygame.typing import ColorLike
 from typing_extensions import deprecated  # added in 3.13
@@ -47,8 +47,8 @@ class Color(Collection[int]):
     def __index__(self) -> int: ...
     def __invert__(self) -> Color: ...
     def __contains__(self, other: int) -> bool: ...  # type: ignore[override]
-    def __getattribute__(self, name: str) -> Union[Color, tuple[int, ...]]: ...
-    def __setattr__(self, name: str, value: Union[Color, tuple[int, ...]]) -> None: ...
+    def __getattribute__(self, name: str) -> Color | tuple[int, ...]: ...
+    def __setattr__(self, name: str, value: Color | tuple[int, ...]) -> None: ...
     @overload
     @classmethod
     def from_cmy(cls, object: tuple[float, float, float], /) -> Color: ...
