@@ -1,4 +1,4 @@
-from typing import Any, ClassVar, Optional, Union, final
+from typing import Any, ClassVar, TypeAlias, final
 
 from pygame.typing import SequenceLike
 from typing_extensions import deprecated  # added in 3.13
@@ -33,21 +33,21 @@ class Event(_GenericEvent):
 class EventType(_GenericEvent):
     pass
 
-_EventTypes = Union[int, SequenceLike[int]]
+_EventTypes: TypeAlias = int | SequenceLike[int]
 
 def pump() -> None: ...
 def get(
-    eventtype: Optional[_EventTypes] = None,
+    eventtype: _EventTypes | None = None,
     pump: bool = True,
-    exclude: Optional[_EventTypes] = None,
+    exclude: _EventTypes | None = None,
 ) -> list[Event]: ...
 def poll() -> Event: ...
 def wait(timeout: int = 0) -> Event: ...
-def peek(eventtype: Optional[_EventTypes] = None, pump: bool = True) -> bool: ...
-def clear(eventtype: Optional[_EventTypes] = None, pump: bool = True) -> None: ...
+def peek(eventtype: _EventTypes | None = None, pump: bool = True) -> bool: ...
+def clear(eventtype: _EventTypes | None = None, pump: bool = True) -> None: ...
 def event_name(type: int, /) -> str: ...
-def set_blocked(type: Optional[_EventTypes], /) -> None: ...
-def set_allowed(type: Optional[_EventTypes], /) -> None: ...
+def set_blocked(type: _EventTypes | None, /) -> None: ...
+def set_allowed(type: _EventTypes | None, /) -> None: ...
 def get_blocked(type: _EventTypes, /) -> bool: ...
 def set_grab(grab: bool, /) -> None: ...
 def get_grab() -> bool: ...
