@@ -1,5 +1,5 @@
 from collections.abc import Callable, Hashable, Iterable
-from typing import Literal, Optional, Union
+from typing import Literal
 
 from pygame.surface import Surface
 from pygame.typing import ColorLike, FileLike
@@ -16,16 +16,16 @@ def get_sdl_ttf_version(linked: bool = True) -> tuple[int, int, int]: ...
 def get_default_font() -> str: ...
 def get_fonts() -> list[str]: ...
 def match_font(
-    name: Union[str, bytes, Iterable[Union[str, bytes]]],
+    name: str | bytes | Iterable[str | bytes],
     bold: Hashable = False,
     italic: Hashable = False,
 ) -> str: ...
 def SysFont(
-    name: Union[str, bytes, Iterable[Union[str, bytes]], None],
+    name: str | bytes | Iterable[str | bytes] | None,
     size: int,
     bold: Hashable = False,
     italic: Hashable = False,
-    constructor: Optional[Callable[[Optional[str], int, bool, bool], Font]] = None,
+    constructor: Callable[[str | None, int, bool, bool], Font] | None = None,
 ) -> Font: ...
 
 class Font:
@@ -57,16 +57,16 @@ class Font:
     def point_size(self) -> int: ...
     @point_size.setter
     def point_size(self, value: int) -> None: ...
-    def __init__(self, filename: Optional[FileLike] = None, size: int = 20) -> None: ...
+    def __init__(self, filename: FileLike | None = None, size: int = 20) -> None: ...
     def render(
         self,
-        text: Union[str, bytes, None],
+        text: str | bytes | None,
         antialias: bool,
         color: ColorLike,
-        bgcolor: Optional[ColorLike] = None,
+        bgcolor: ColorLike | None = None,
         wraplength: int = 0,
     ) -> Surface: ...
-    def size(self, text: Union[str, bytes], /) -> tuple[int, int]: ...
+    def size(self, text: str | bytes, /) -> tuple[int, int]: ...
     def set_underline(self, value: bool, /) -> None: ...
     def get_underline(self) -> bool: ...
     def set_strikethrough(self, value: bool, /) -> None: ...
@@ -74,9 +74,7 @@ class Font:
     def set_bold(self, value: bool, /) -> None: ...
     def get_bold(self) -> bool: ...
     def set_italic(self, value: bool, /) -> None: ...
-    def metrics(
-        self, text: Union[str, bytes], /
-    ) -> list[tuple[int, int, int, int, int]]: ...
+    def metrics(self, text: str | bytes, /) -> list[tuple[int, int, int, int, int]]: ...
     def get_italic(self) -> bool: ...
     def get_linesize(self) -> int: ...
     def set_linesize(self, linesize: int, /) -> None: ...
