@@ -62,9 +62,8 @@ following formats.
 .. versionaddedold:: 1.8 Saving PNG and JPEG files.
 """
 
-from typing import Literal, Optional, Union
+from typing import Literal, TypeAlias
 
-from pygame.bufferproxy import BufferProxy
 from pygame.surface import Surface
 from pygame.typing import FileLike, IntPoint, Point
 from typing_extensions import (
@@ -72,11 +71,15 @@ from typing_extensions import (
     deprecated,  # added in 3.13
 )
 
-_from_buffer_format = Literal["P", "RGB", "BGR", "BGRA", "RGBX", "RGBA", "ARGB"]
-_to_bytes_format = Literal[
+_FromBufferFormat: TypeAlias = Literal[
+    "P", "RGB", "BGR", "BGRA", "RGBX", "RGBA", "ARGB"
+]
+_ToBytesFormat: TypeAlias = Literal[
     "P", "RGB", "RGBX", "RGBA", "ARGB", "BGRA", "ABGR", "RGBA_PREMULT", "ARGB_PREMULT"
 ]
-_from_bytes_format = Literal["P", "RGB", "RGBX", "RGBA", "ARGB", "BGRA", "ABGR"]
+_FromBytesFormat: TypeAlias = Literal[
+    "P", "RGB", "RGBX", "RGBA", "ARGB", "BGRA", "ABGR"
+]
 
 def load(file: FileLike, namehint: str = "") -> Surface:
     """Load new image from a file (or file-like object).
@@ -177,7 +180,7 @@ def save(surface: Surface, file: FileLike, namehint: str = "") -> None:
     .. versionchanged:: 2.2.0 Now supports keyword arguments.
     """
 
-def get_sdl_image_version(linked: bool = True) -> Optional[tuple[int, int, int]]:
+def get_sdl_image_version(linked: bool = True) -> tuple[int, int, int] | None:
     """Get version number of the SDL_Image library being used.
 
     If pygame is built with extended image formats, then this function will
@@ -206,7 +209,7 @@ def get_extended() -> bool:
 @deprecated("since 2.3.0. Use `pygame.image.tobytes` instead")
 def tostring(
     surface: Surface,
-    format: _to_bytes_format,
+    format: _ToBytesFormat,
     flipped: bool = False,
     pitch: int = -1,
 ) -> bytes:
@@ -220,7 +223,7 @@ def tostring(
 
 def tobytes(
     surface: Surface,
-    format: _to_bytes_format,
+    format: _ToBytesFormat,
     flipped: bool = False,
     pitch: int = -1,
 ) -> bytes:
@@ -275,7 +278,7 @@ def tobytes(
 def fromstring(
     bytes: bytes,
     size: IntPoint,
-    format: _from_bytes_format,
+    format: _FromBytesFormat,
     flipped: bool = False,
     pitch: int = -1,
 ) -> Surface:
@@ -290,7 +293,7 @@ def fromstring(
 def frombytes(
     bytes: bytes,
     size: IntPoint,
-    format: _from_bytes_format,
+    format: _FromBytesFormat,
     flipped: bool = False,
     pitch: int = -1,
 ) -> Surface:
@@ -324,7 +327,7 @@ def frombytes(
 def frombuffer(
     buffer: Buffer,
     size: IntPoint,
-    format: _from_buffer_format,
+    format: _FromBufferFormat,
     pitch: int = -1,
 ) -> Surface:
     """Create a new Surface that shares data inside a bytes buffer.
