@@ -4,10 +4,8 @@ from typing import (
     ClassVar,
     Generic,
     Literal,
-    Optional,
     SupportsIndex,
     TypeVar,
-    Union,
     final,
     overload,
 )
@@ -29,20 +27,18 @@ class _GenericVector(Collection[float]):
     @overload
     def __setitem__(self, key: int, value: float) -> None: ...
     @overload
-    def __setitem__(
-        self, key: slice, value: Union[SequenceLike[float], _TVec]
-    ) -> None: ...
+    def __setitem__(self, key: slice, value: SequenceLike[float] | _TVec) -> None: ...
     @overload
     def __getitem__(self, i: SupportsIndex) -> float: ...
     @overload
     def __getitem__(self, s: slice) -> list[float]: ...
     def __iter__(self) -> VectorIterator: ...
-    def __add__(self: _TVec, other: Union[SequenceLike[float], _TVec]) -> _TVec: ...
-    def __radd__(self: _TVec, other: Union[SequenceLike[float], _TVec]) -> _TVec: ...
-    def __sub__(self: _TVec, other: Union[SequenceLike[float], _TVec]) -> _TVec: ...
-    def __rsub__(self: _TVec, other: Union[SequenceLike[float], _TVec]) -> _TVec: ...
+    def __add__(self: _TVec, other: SequenceLike[float] | _TVec) -> _TVec: ...
+    def __radd__(self: _TVec, other: SequenceLike[float] | _TVec) -> _TVec: ...
+    def __sub__(self: _TVec, other: SequenceLike[float] | _TVec) -> _TVec: ...
+    def __rsub__(self: _TVec, other: SequenceLike[float] | _TVec) -> _TVec: ...
     @overload
-    def __mul__(self: _TVec, other: Union[SequenceLike[float], _TVec]) -> float: ...
+    def __mul__(self: _TVec, other: SequenceLike[float] | _TVec) -> float: ...
     @overload
     def __mul__(self: _TVec, other: float) -> _TVec: ...
     def __rmul__(self: _TVec, other: float) -> _TVec: ...
@@ -52,17 +48,17 @@ class _GenericVector(Collection[float]):
     def __neg__(self: _TVec) -> _TVec: ...
     def __pos__(self: _TVec) -> _TVec: ...
     def __bool__(self) -> bool: ...
-    def __iadd__(self: _TVec, other: Union[SequenceLike[float], _TVec]) -> _TVec: ...
-    def __isub__(self: _TVec, other: Union[SequenceLike[float], _TVec]) -> _TVec: ...
+    def __iadd__(self: _TVec, other: SequenceLike[float] | _TVec) -> _TVec: ...
+    def __isub__(self: _TVec, other: SequenceLike[float] | _TVec) -> _TVec: ...
     @overload
-    def __imul__(self: _TVec, other: Union[SequenceLike[float], _TVec]) -> float: ...
+    def __imul__(self: _TVec, other: SequenceLike[float] | _TVec) -> float: ...
     @overload
     def __imul__(self: _TVec, other: float) -> _TVec: ...
     def __copy__(self: _TVec) -> _TVec: ...
     def copy(self: _TVec) -> _TVec: ...
     def __safe_for_unpickling__(self) -> Literal[True]: ...
     def __contains__(self, other: float) -> bool: ...  # type: ignore[override]
-    def dot(self: _TVec, other: Union[SequenceLike[float], _TVec], /) -> float: ...
+    def dot(self: _TVec, other: SequenceLike[float] | _TVec, /) -> float: ...
     def magnitude(self) -> float: ...
     def magnitude_squared(self) -> float: ...
     def length(self) -> float: ...
@@ -71,32 +67,28 @@ class _GenericVector(Collection[float]):
     def normalize_ip(self) -> None: ...
     def is_normalized(self) -> bool: ...
     def scale_to_length(self, value: float, /) -> None: ...
-    def reflect(self: _TVec, other: Union[SequenceLike[float], _TVec], /) -> _TVec: ...
-    def reflect_ip(
-        self: _TVec, other: Union[SequenceLike[float], _TVec], /
-    ) -> None: ...
-    def distance_to(
-        self: _TVec, other: Union[SequenceLike[float], _TVec], /
-    ) -> float: ...
+    def reflect(self: _TVec, other: SequenceLike[float] | _TVec, /) -> _TVec: ...
+    def reflect_ip(self: _TVec, other: SequenceLike[float] | _TVec, /) -> None: ...
+    def distance_to(self: _TVec, other: SequenceLike[float] | _TVec, /) -> float: ...
     def distance_squared_to(
-        self: _TVec, other: Union[SequenceLike[float], _TVec], /
+        self: _TVec, other: SequenceLike[float] | _TVec, /
     ) -> float: ...
     def lerp(
-        self: _TVec, other: Union[SequenceLike[float], _TVec], value: float, /
+        self: _TVec, other: SequenceLike[float] | _TVec, value: float, /
     ) -> _TVec: ...
     def slerp(
-        self: _TVec, other: Union[SequenceLike[float], _TVec], value: float, /
+        self: _TVec, other: SequenceLike[float] | _TVec, value: float, /
     ) -> _TVec: ...
     def smoothstep(
-        self: _TVec, other: Union[SequenceLike[float], _TVec], value: float, /
+        self: _TVec, other: SequenceLike[float] | _TVec, value: float, /
     ) -> _TVec: ...
     def elementwise(self: _TVec) -> VectorElementwiseProxy[_TVec]: ...
-    def angle_to(self: _TVec, other: Union[SequenceLike[float], _TVec], /) -> float: ...
+    def angle_to(self: _TVec, other: SequenceLike[float] | _TVec, /) -> float: ...
     def move_towards(
-        self: _TVec, target: Union[SequenceLike[float], _TVec], max_distance: float, /
+        self: _TVec, target: SequenceLike[float] | _TVec, max_distance: float, /
     ) -> _TVec: ...
     def move_towards_ip(
-        self: _TVec, target: Union[SequenceLike[float], _TVec], max_distance: float, /
+        self: _TVec, target: SequenceLike[float] | _TVec, max_distance: float, /
     ) -> None: ...
     @overload
     def clamp_magnitude(self: _TVec, max_length: float, /) -> _TVec: ...
@@ -108,8 +100,8 @@ class _GenericVector(Collection[float]):
     def clamp_magnitude_ip(self, max_length: float, /) -> None: ...
     @overload
     def clamp_magnitude_ip(self, min_length: float, max_length: float, /) -> None: ...
-    def project(self: _TVec, other: Union[SequenceLike[float], _TVec], /) -> _TVec: ...
-    def __round__(self: _TVec, ndigits: Optional[int] = None, /) -> _TVec: ...
+    def project(self: _TVec, other: SequenceLike[float] | _TVec, /) -> _TVec: ...
+    def __round__(self: _TVec, ndigits: int | None = None, /) -> _TVec: ...
 
 # VectorElementwiseProxy is a generic, it can be an elementwiseproxy object for
 # Vector2, Vector3 and vector subclass objects
@@ -117,79 +109,79 @@ class _GenericVector(Collection[float]):
 class VectorElementwiseProxy(Generic[_TVec]):
     def __add__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __radd__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __sub__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __rsub__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __mul__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __rmul__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __truediv__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __rtruediv__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __floordiv__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __rfloordiv__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __mod__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __rmod__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> _TVec: ...
     def __pow__(
         self,
-        power: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        power: float | _TVec | VectorElementwiseProxy[_TVec],
         mod: None = None,
     ) -> _TVec: ...
     def __rpow__(
         self,
-        power: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        power: float | _TVec | VectorElementwiseProxy[_TVec],
         mod: None = None,
     ) -> _TVec: ...
     def __eq__(self, other: Any) -> bool: ...
     def __ne__(self, other: Any) -> bool: ...
     def __gt__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> bool: ...
     def __lt__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> bool: ...
     def __ge__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> bool: ...
     def __le__(
         self,
-        other: Union[float, _TVec, VectorElementwiseProxy[_TVec]],
+        other: float | _TVec | VectorElementwiseProxy[_TVec],
     ) -> bool: ...
     def __abs__(self) -> _TVec: ...
     def __neg__(self) -> _TVec: ...
@@ -216,7 +208,7 @@ class Vector2(_GenericVector):
     @overload
     def __init__(
         self: _TVec,
-        x: Union[str, float, SequenceLike[float], _TVec] = 0,
+        x: str | float | SequenceLike[float] | _TVec = 0,
     ) -> None: ...
     @overload
     def __init__(self, x: float, y: float) -> None: ...
@@ -227,13 +219,13 @@ class Vector2(_GenericVector):
     def rotate_rad_ip(self, angle: float, /) -> None: ...
     @deprecated("since 2.1.1. Use `pygame.Vector2.rotate_rad_ip` instead")
     def rotate_ip_rad(self, angle: float, /) -> None: ...
-    def cross(self: _TVec, other: Union[SequenceLike[float], _TVec], /) -> float: ...
+    def cross(self: _TVec, other: SequenceLike[float] | _TVec, /) -> float: ...
     def as_polar(self) -> tuple[float, float]: ...
     def from_polar(self, polar_value: SequenceLike[float], /) -> None: ...
     @overload
     def update(
         self: _TVec,
-        x: Union[str, float, SequenceLike[float], _TVec] = 0,
+        x: str | float | SequenceLike[float] | _TVec = 0,
     ) -> None: ...
     @overload
     def update(self, x: float = 0, y: float = 0) -> None: ...
@@ -281,27 +273,27 @@ class Vector3(_GenericVector):
     @overload
     def __init__(
         self: _TVec,
-        x: Union[str, float, SequenceLike[float], _TVec] = 0,
+        x: str | float | SequenceLike[float] | _TVec = 0,
     ) -> None: ...
     @overload
     def __init__(self, x: float, y: float, z: float) -> None: ...
     def __reduce__(self: _TVec) -> tuple[type[_TVec], tuple[float, float, float]]: ...
-    def cross(self: _TVec, other: Union[SequenceLike[float], _TVec], /) -> _TVec: ...
+    def cross(self: _TVec, other: SequenceLike[float] | _TVec, /) -> _TVec: ...
     def rotate(
-        self: _TVec, angle: float, axis: Union[SequenceLike[float], _TVec], /
+        self: _TVec, angle: float, axis: SequenceLike[float] | _TVec, /
     ) -> _TVec: ...
     def rotate_rad(
-        self: _TVec, angle: float, axis: Union[SequenceLike[float], _TVec], /
+        self: _TVec, angle: float, axis: SequenceLike[float] | _TVec, /
     ) -> _TVec: ...
     def rotate_ip(
-        self: _TVec, angle: float, axis: Union[SequenceLike[float], _TVec], /
+        self: _TVec, angle: float, axis: SequenceLike[float] | _TVec, /
     ) -> None: ...
     def rotate_rad_ip(
-        self: _TVec, angle: float, axis: Union[SequenceLike[float], _TVec], /
+        self: _TVec, angle: float, axis: SequenceLike[float] | _TVec, /
     ) -> None: ...
     @deprecated("since 2.1.1. Use `pygame.Vector3.rotate_rad_ip` instead")
     def rotate_ip_rad(
-        self: _TVec, angle: float, axis: Union[SequenceLike[float], _TVec], /
+        self: _TVec, angle: float, axis: SequenceLike[float] | _TVec, /
     ) -> None: ...
     def rotate_x(self: _TVec, angle: float, /) -> _TVec: ...
     def rotate_x_rad(self: _TVec, angle: float, /) -> _TVec: ...
@@ -326,7 +318,7 @@ class Vector3(_GenericVector):
     @overload
     def update(
         self: _TVec,
-        x: Union[str, float, SequenceLike[float], _TVec] = 0,
+        x: str | float | SequenceLike[float] | _TVec = 0,
     ) -> None: ...
     @overload
     def update(self, x: int, y: int, z: int) -> None: ...
