@@ -1,5 +1,4 @@
 from collections.abc import Iterable
-from typing import Union
 
 from pygame.event import Event
 from pygame.typing import SequenceLike
@@ -18,7 +17,7 @@ def get_default_input_id() -> int: ...
 def get_default_output_id() -> int: ...
 def get_device_info(an_id: int) -> tuple[str, str, int, int, int]: ...
 def midis2events(
-    midis: Iterable[SequenceLike[Union[SequenceLike[int], int]]], device_id: int
+    midis: Iterable[SequenceLike[SequenceLike[int] | int]], device_id: int
 ) -> list[Event]: ...
 def time() -> int: ...
 def frequency_to_midi(frequency: float) -> int: ...
@@ -30,7 +29,7 @@ class Input:
     def __init__(self, device_id: int, buffer_size: int = 4096) -> None: ...
     def close(self) -> None: ...
     def poll(self) -> bool: ...
-    def read(self, num_events: int) -> list[list[Union[list[int], int]]]: ...
+    def read(self, num_events: int) -> list[list[list[int] | int]]: ...
 
 class Output:
     device_id: int
@@ -46,11 +45,9 @@ class Output:
     def note_on(self, note: int, velocity: int, channel: int = 0) -> None: ...
     def set_instrument(self, instrument_id: int, channel: int = 0) -> None: ...
     def pitch_bend(self, value: int = 0, channel: int = 0) -> None: ...
-    def write(
-        self, data: Iterable[SequenceLike[Union[SequenceLike[int], int]]]
-    ) -> None: ...
+    def write(self, data: Iterable[SequenceLike[SequenceLike[int] | int]]) -> None: ...
     def write_short(self, status: int, data1: int = 0, data2: int = 0) -> None: ...
-    def write_sys_ex(self, when: int, msg: Union[list[int], str]) -> None: ...
+    def write_sys_ex(self, when: int, msg: list[int] | str) -> None: ...
 
 # keep in sync with midi.py
 __all__ = [
