@@ -1,5 +1,5 @@
 import sys
-from typing import Any, Optional, Union, overload
+from typing import Any, overload
 
 from pygame.event import Event
 from pygame.typing import FileLike
@@ -19,7 +19,7 @@ def init(
     size: int = -16,
     channels: int = 2,
     buffer: int = 512,
-    devicename: Optional[str] = None,
+    devicename: str | None = None,
     allowedchanges: int = 5,
 ) -> None: ...
 def pre_init(
@@ -27,7 +27,7 @@ def pre_init(
     size: int = -16,
     channels: int = 2,
     buffer: int = 512,
-    devicename: Optional[str] = None,
+    devicename: str | None = None,
     allowedchanges: int = 5,
 ) -> None: ...
 def quit() -> None: ...
@@ -41,8 +41,8 @@ def set_num_channels(count: int, /) -> None: ...
 def get_num_channels() -> int: ...
 def set_reserved(count: int, /) -> int: ...
 def find_channel(force: bool = False) -> Channel: ...
-def set_soundfont(paths: Optional[str] = None, /) -> None: ...
-def get_soundfont() -> Optional[str]: ...
+def set_soundfont(paths: str | None = None, /) -> None: ...
+def get_soundfont() -> str | None: ...
 def get_busy() -> bool: ...
 def get_sdl_mixer_version(linked: bool = True) -> tuple[int, int, int]: ...
 
@@ -101,7 +101,7 @@ class Channel:
     def get_busy(self) -> bool: ...
     def get_sound(self) -> Sound: ...
     def get_queue(self) -> Sound: ...
-    def set_endevent(self, type: Union[int, Event] = 0, /) -> None: ...
+    def set_endevent(self, type: int | Event = 0, /) -> None: ...
     def get_endevent(self) -> int: ...
 
 @deprecated("Use `Sound` instead (SoundType is an old alias)")
