@@ -3,10 +3,8 @@ from collections.abc import Callable, Collection, Iterator
 from typing import (
     ClassVar,
     Literal,
-    Optional,
     SupportsIndex,
     TypeVar,
-    Union,
     overload,
 )
 
@@ -18,12 +16,7 @@ if sys.version_info >= (3, 11):
 else:
     from typing_extensions import Self
 
-# 'ellipsis' existed in typeshed pre 3.10, now we use EllipsisType which is
-# the modern standard library equivalent.
-if sys.version_info >= (3, 10):
-    from types import EllipsisType
-else:
-    EllipsisType = ellipsis
+from types import EllipsisType
 
 _N = TypeVar("_N", int, float)
 _K = TypeVar("_K")
@@ -145,12 +138,12 @@ class _GenericRect(Collection[_N]):
     @overload
     def __getitem__(self, i: SupportsIndex) -> _N: ...
     @overload
-    def __getitem__(self, s: Union[slice, EllipsisType]) -> list[_N]: ...
+    def __getitem__(self, s: slice | EllipsisType) -> list[_N]: ...
     @overload
     def __setitem__(self, key: int, value: float) -> None: ...
     @overload
     def __setitem__(
-        self, key: Union[slice, EllipsisType], value: Union[float, RectLike]
+        self, key: slice | EllipsisType, value: float | RectLike
     ) -> None: ...
     def __copy__(self) -> Self: ...
     def copy(self) -> Self: ...
@@ -162,7 +155,7 @@ class _GenericRect(Collection[_N]):
     def move_ip(self, x: float, y: float, /) -> None: ...
     @overload
     def move_ip(self, move_by: Point, /) -> None: ...
-    def move_to(self, **kwargs: Union[float, Point]) -> Self: ...
+    def move_to(self, **kwargs: float | Point) -> Self: ...
     @overload
     def inflate(self, x: float, y: float, /) -> Self: ...
     @overload
@@ -212,15 +205,15 @@ class _GenericRect(Collection[_N]):
     @overload
     def clipline(
         self, x1: float, x2: float, x3: float, x4: float, /
-    ) -> Union[tuple[tuple[_N, _N], tuple[_N, _N]], tuple[()]]: ...
+    ) -> tuple[tuple[_N, _N], tuple[_N, _N]] | tuple[()]: ...
     @overload
     def clipline(
         self, first_point: Point, second_point: Point, /
-    ) -> Union[tuple[tuple[_N, _N], tuple[_N, _N]], tuple[()]]: ...
+    ) -> tuple[tuple[_N, _N], tuple[_N, _N]] | tuple[()]: ...
     @overload
     def clipline(
         self, rect_arg: RectLike, /
-    ) -> Union[tuple[tuple[_N, _N], tuple[_N, _N]], tuple[()]]: ...
+    ) -> tuple[tuple[_N, _N], tuple[_N, _N]] | tuple[()]: ...
     @overload
     def union(self, rect: RectLike, /) -> Self: ...
     @overload
@@ -248,7 +241,7 @@ class _GenericRect(Collection[_N]):
     @overload
     def fit(self, left: float, top: float, width: float, height: float, /) -> Self: ...
     def normalize(self) -> None: ...
-    def __contains__(self, rect: Union[RectLike, _N], /) -> bool: ...  # type: ignore[override]
+    def __contains__(self, rect: RectLike | _N, /) -> bool: ...  # type: ignore[override]
     @overload
     def contains(self, rect: RectLike, /) -> bool: ...
     @overload
@@ -276,21 +269,21 @@ class _GenericRect(Collection[_N]):
         self, rect_list: SequenceLike[_RectTypeCompatible_co], /
     ) -> list[int]: ...
     def collideobjectsall(
-        self, objects: SequenceLike[_T], key: Optional[Callable[[_T], RectLike]] = None
+        self, objects: SequenceLike[_T], key: Callable[[_T], RectLike] | None = None
     ) -> list[_T]: ...
     def collideobjects(
-        self, objects: SequenceLike[_T], key: Optional[Callable[[_T], RectLike]] = None
-    ) -> Optional[_T]: ...
+        self, objects: SequenceLike[_T], key: Callable[[_T], RectLike] | None = None
+    ) -> _T | None: ...
     @overload
     def collidedict(
         self,
         rect_dict: dict[_RectTypeCompatible_co, _V],
         values: Literal[False] = False,
-    ) -> Optional[tuple[_RectTypeCompatible_co, _V]]: ...
+    ) -> tuple[_RectTypeCompatible_co, _V] | None: ...
     @overload
     def collidedict(
         self, rect_dict: dict[_K, _RectTypeCompatible_co], values: Literal[True]
-    ) -> Optional[tuple[_K, _RectTypeCompatible_co]]: ...
+    ) -> tuple[_K, _RectTypeCompatible_co] | None: ...
     @overload
     def collidedictall(
         self,
