@@ -1,5 +1,3 @@
-from typing import Optional
-
 from typing_extensions import (
     Buffer,  # added in 3.12,
     deprecated,  # added in 3.13
@@ -10,7 +8,7 @@ def init() -> None: ...
 @deprecated("since 2.2.0. Use the new API instead, which doesn't require scrap init")
 def get_init() -> bool: ...
 @deprecated("since 2.2.0. Use the new API instead: `pygame.scrap.get_text`")
-def get(data_type: str, /) -> Optional[bytes]: ...
+def get(data_type: str, /) -> bytes | None: ...
 @deprecated("since 2.2.0. Use the new API instead, which only supports strings")
 def get_types() -> list[str]: ...
 @deprecated("since 2.2.0. Use the new API instead: `pygame.scrap.put_text`")
