@@ -15,15 +15,15 @@ __all__ = [
 from abc import abstractmethod
 from collections.abc import Callable
 from os import PathLike as _PathProtocol
-from typing import IO, Protocol, TypeVar, Union
+from typing import IO, Protocol, TypeAlias, TypeVar
 
 from pygame.color import Color
 from pygame.rect import FRect, Rect
 
 # For functions that take a file name
-_PathLike = Union[str, bytes, _PathProtocol[str], _PathProtocol[bytes]]
+_PathLike: TypeAlias = str | bytes | _PathProtocol[str] | _PathProtocol[bytes]
 # Most pygame functions that take a file argument should be able to handle a FileLike type
-FileLike = Union[_PathLike, IO[bytes], IO[str]]
+FileLike: TypeAlias = _PathLike | IO[bytes] | IO[str]
 
 _T_co = TypeVar("_T_co", covariant=True)
 
@@ -42,24 +42,24 @@ class SequenceLike(Protocol[_T_co]):
 # Modify typehints when it is possible to annotate sizes
 
 # Pygame handles float without errors in most cases where a point is expected,
-# usually rounding to int. Also, 'Union[int, float] == float'
-Point = SequenceLike[float]
+# usually rounding to int. Also, 'int | float == float'
+Point: TypeAlias = SequenceLike[float]
 # This is used where ints are strictly required
-IntPoint = SequenceLike[int]
+IntPoint: TypeAlias = SequenceLike[int]
 
-ColorLike = Union[Color, SequenceLike[int], str, int]
+ColorLike: TypeAlias = Color | SequenceLike[int] | str | int
 
 
 class _HasRectAttribute(Protocol):
     # An object that has a rect attribute that is either a rect, or a function
     # that returns a rect conforms to the rect protocol
     @property
-    def rect(self) -> Union["RectLike", Callable[[], "RectLike"]]: ...
+    def rect(self) -> "RectLike | Callable[[], RectLike]": ...
 
 
-RectLike = Union[
-    Rect, FRect, SequenceLike[float], SequenceLike[Point], _HasRectAttribute
-]
+RectLike: TypeAlias = (
+    Rect | FRect | SequenceLike[float] | SequenceLike[Point] | _HasRectAttribute
+)
 
 
 # cleanup namespace
@@ -70,7 +70,6 @@ del (
     FRect,
     IO,
     Callable,
-    Union,
     TypeVar,
     Protocol,
 )
