@@ -1,12 +1,10 @@
-from typing import Optional, Union
-
 from pygame.locals import WINDOWPOS_UNDEFINED
 from pygame.rect import Rect
 from pygame.surface import Surface
 from pygame.typing import Point, RectLike
 from typing_extensions import deprecated  # added in 3.13
 
-def get_grabbed_window() -> Optional[Window]: ...
+def get_grabbed_window() -> Window | None: ...
 
 class Window:
     """Pygame object that represents a window.
@@ -93,7 +91,7 @@ class Window:
         self,
         title: str = "pygame window",
         size: Point = (640, 480),
-        position: Union[int, Point] = WINDOWPOS_UNDEFINED,
+        position: int | Point = WINDOWPOS_UNDEFINED,
         *,
         fullscreen: bool = ...,
         fullscreen_desktop: bool = ...,
@@ -233,7 +231,7 @@ class Window:
         """Get the unique window ID (**read-only**)."""
 
     @property
-    def mouse_rect(self) -> Optional[Rect]:
+    def mouse_rect(self) -> Rect | None:
         """Get or set the mouse confinement rectangle of the window.
 
         Setting this attribute to a rect-like object confines the
@@ -248,7 +246,7 @@ class Window:
         """
 
     @mouse_rect.setter
-    def mouse_rect(self, value: Optional[RectLike]) -> None: ...
+    def mouse_rect(self, value: RectLike | None) -> None: ...
     @property
     def size(self) -> tuple[int, int]:
         """Get or set the window size in pixels."""
@@ -305,7 +303,7 @@ class Window:
         """
 
     @position.setter
-    def position(self, value: Union[int, Point]) -> None: ...
+    def position(self, value: int | Point) -> None: ...
     opacity: float
     """Get or set the window opacity, between 0.0 (fully transparent) and 1.0 (fully opaque).
 
