@@ -1,7 +1,7 @@
 /* Auto generated file: with make_docs.py .  Docs go in docs/reST/ref/ . */
 #define DOC_SURFACE "Surface(size, flags=0, depth=0, masks=None) -> Surface\nSurface(size, flags=0, surface=...) -> Surface\nPygame object for representing images."
 #define DOC_SURFACE_BLIT "blit(source, dest=(0, 0), area=None, special_flags=0) -> Rect\nDraw another Surface onto this one."
-#define DOC_SURFACE_BLITS "blits(blit_sequence, doreturn=1) -> Union[list[Rect], None]\nDraw many Surfaces onto this Surface at their corresponding location."
+#define DOC_SURFACE_BLITS "blits(blit_sequence, doreturn=1) -> list[Rect] | None\nDraw many Surfaces onto this Surface at their corresponding location."
 #define DOC_SURFACE_FBLITS "fblits(blit_sequence, special_flags=0, /) -> None\nDraw many Surfaces onto this Surface at their corresponding location and with the same special_flags."
 #define DOC_SURFACE_CONVERT "convert(surface, /) -> Surface\nconvert(depth, flags=0, /) -> Surface\nconvert(masks, flags=0, /) -> Surface\nconvert() -> Surface\nChange the pixel format of a Surface."
 #define DOC_SURFACE_CONVERTALPHA "convert_alpha() -> Surface\nChange the pixel format of a Surface including per pixel alphas."
@@ -9,9 +9,9 @@
 #define DOC_SURFACE_FILL "fill(color, rect=None, special_flags=0) -> Rect\nFill Surface with a solid color."
 #define DOC_SURFACE_SCROLL "scroll(dx=0, dy=0, scroll_flag=0, /) -> None\nShift the Surface pixels in place."
 #define DOC_SURFACE_SETCOLORKEY "set_colorkey(color, flags=0, /) -> None\nset_colorkey(color, /) -> None\nSet the transparent colorkey."
-#define DOC_SURFACE_GETCOLORKEY "get_colorkey() -> Optional[tuple[int, int, int, int]]\nGet the current transparent colorkey."
+#define DOC_SURFACE_GETCOLORKEY "get_colorkey() -> tuple[int, int, int, int] | None\nGet the current transparent colorkey."
 #define DOC_SURFACE_SETALPHA "set_alpha(value, flags=0, /) -> None\nset_alpha(value, /) -> None\nSet the alpha value for the full Surface."
-#define DOC_SURFACE_GETALPHA "get_alpha() -> Optional[int]\nGet the current Surface transparency value."
+#define DOC_SURFACE_GETALPHA "get_alpha() -> int | None\nGet the current Surface transparency value."
 #define DOC_SURFACE_LOCK "lock() -> None\nLock the Surface memory for pixel access."
 #define DOC_SURFACE_UNLOCK "unlock() -> None\nUnlock the Surface memory from pixel access."
 #define DOC_SURFACE_MUSTLOCK "mustlock() -> bool\nTest if the Surface requires locking."
